+++ MQ2Mono.cpp
@@ -14,53 +14,11 @@
 #include <imgui/imgui.h>
 #include <imgui/imgui_internal.h>
 #include <mq/imgui/Widgets.h>
+#include "MQ2MonoImGui.h"
+#include "MQ2MonoShared.h"
 PreSetup("MQ2Mono");
 
-// Forward declarations for new ImGui wrappers (tables, colors)
-bool mono_ImGUI_BeginTable(MonoString* id, int columns, int flags, float outer_width);
-void mono_ImGUI_EndTable();
-void mono_ImGUI_TableSetupColumn(MonoString* label, int flags, float init_width);
-void mono_ImGUI_TableHeadersRow();
-void mono_ImGUI_TableNextRow();
-bool mono_ImGUI_TableNextColumn();
-void mono_ImGUI_SameLineEx(float offset_from_start_x, float spacing);
-void mono_ImGUI_TextColored(float r, float g, float b, float a, MonoString* text);
-void mono_ImGUI_PushStyleColor(int which, float r, float g, float b, float a);
-void mono_ImGUI_PopStyleColor(int count);
-// New: text wrapping and window size constraints
-void mono_ImGUI_TextWrapped(MonoString* text);
-// Safe text without printf formatting
-void mono_ImGUI_TextUnformatted(MonoString* text);
-void mono_ImGUI_PushTextWrapPos(float wrap_local_pos_x);
-void mono_ImGUI_PopTextWrapPos();
-void mono_ImGUI_SetNextWindowSizeConstraints(float min_w, float min_h, float max_w, float max_h);
-// New ImGui wrappers for floating UI
-void mono_ImGUI_SetNextWindowBgAlpha(float alpha);
-void mono_ImGUI_SetNextWindowSize(float width, float height);
-bool mono_ImGUI_ButtonEx(MonoString* name, float width, float height);
-bool mono_ImGUI_IsWindowHovered();
-bool mono_ImGUI_IsMouseClicked(int button);
-bool mono_ImGUI_SmallButton(MonoString* name);
-// New: Sliders
-bool mono_ImGUI_SliderInt(MonoString* id, int* value, int min, int max);
-bool mono_ImGUI_SliderDouble(MonoString* id, double* value, double min, double max, MonoString* fmt);
-// TreeNode helpers
-bool mono_ImGUI_TreeNode(MonoString* label);
-bool mono_ImGUI_TreeNodeEx(MonoString* label, int flags);
-void mono_ImGUI_TreePop();
-bool mono_ImGUI_CollapsingHeader(MonoString* label, int flags);
-// Tooltips and hover detection
-bool mono_ImGUI_IsItemHovered();
-void mono_ImGUI_BeginTooltip();
-void mono_ImGUI_EndTooltip();
-// Image display
-void mono_ImGUI_Image(void* textureId, float width, float height);
-// Draw spell icon helpers
-void mono_ImGUI_DrawSpellIconByIconIndex(int iconIndex, float size);
-void mono_ImGUI_DrawSpellIconBySpellID(int spellId, float size);
-// Texture creation from raw data
-void* mono_CreateTextureFromData(const uint8_t* data, int width, int height, int channels);
-void mono_DestroyTexture(void* textureId);
+// ImGui wrappers moved to MQ2MonoImGui.h / MQ2MonoImGui.cpp
 PLUGIN_VERSION(0.34);
 
 /**
@@ -83,51 +41,6 @@ PLUGIN_VERSION(0.34);
  bool mono_AddCommand(MonoString* string);
  void mono_ClearCommands();
  void mono_RemoveCommand(MonoString* text);
- //IMGUI calls, not in use really was a test
- bool mono_ImGUI_Begin(MonoString* name, int flags);
- bool mono_ImGUI_Button(MonoString* name);
- bool mono_ImGUI_ButtonEx(MonoString* name, float width, float height);
- bool mono_ImGUI_SmallButton(MonoString* name);
- void mono_ImGUI_End();
- boolean mono_ImGUI_Begin_OpenFlagGet(MonoString* name);
- void mono_ImGUI_Begin_OpenFlagSet(MonoString* name,bool open);
- // Extended ImGui wrappers
- void mono_ImGUI_Text(MonoString* text);
- void mono_ImGUI_Separator();
- void mono_ImGUI_SameLine();
- bool mono_ImGUI_Checkbox(MonoString* name, bool defaultValue);
- // Tab bar wrappers
- bool mono_ImGUI_BeginTabBar(MonoString* name);
- void mono_ImGUI_EndTabBar();
- bool mono_ImGUI_BeginTabItem(MonoString* label);
- void mono_ImGUI_EndTabItem();
-  // Child/selection/layout helpers
-  bool mono_ImGUI_BeginChild(MonoString* id, float width, float height, bool border);
-  void mono_ImGUI_EndChild();
-  bool mono_ImGUI_Selectable(MonoString* label, bool selected);
-  float mono_ImGUI_GetContentRegionAvailX();
-  float mono_ImGUI_GetContentRegionAvailY();
-  // Text input helpers
-  bool mono_ImGUI_InputText(MonoString* id, MonoString* initial);
-  MonoString* mono_ImGUI_InputText_Get(MonoString* id);
-  void mono_ImGUI_SetNextItemWidth(float width);
-  // Window control helpers
-  void mono_ImGUI_SetNextWindowBgAlpha(float alpha);
-  void mono_ImGUI_SetNextWindowSize(float width, float height);
-  bool mono_ImGUI_IsWindowHovered();
-  bool mono_ImGUI_IsMouseClicked(int button);
-  // Combo helpers
-  bool mono_ImGUI_BeginCombo(MonoString* label, MonoString* preview, int flags);
-  void mono_ImGUI_EndCombo();
-  // Alignment helper
-  bool mono_ImGUI_RightAlignButton(MonoString* name);
-  // Context menu / popup helpers
-  bool mono_ImGUI_BeginPopupContextItem(MonoString* id, int flags);
-  bool mono_ImGUI_BeginPopupContextWindow(MonoString* id, int flags);
-  void mono_ImGUI_EndPopup();
-  bool mono_ImGUI_MenuItem(MonoString* label);
-  //end temp methods
- 
  bool InitAppDomain(std::string appDomainName);
  bool UnloadAppDomain(std::string appDomainName, bool updateCollections);
  void UnloadAllAppDomains();
@@ -155,100 +68,10 @@ PLUGIN_VERSION(0.34);
  /// Main data structure that has information on each individual app domain that we create and informatoin
  /// we need to keep track of.
  /// </summary>
-struct monoAppDomainInfo
-{
-	 std::string m_appDomainName;
-	 //app domain we have created for e3
-	 MonoDomain* m_appDomain=nullptr;
-	 //core.dll information so we can bind to it
-	 MonoAssembly* m_csharpAssembly = nullptr;
-	 MonoImage* m_coreAssemblyImage = nullptr;
-	 MonoClass* m_classInfo = nullptr;
-	 MonoObject* m_classInstance = nullptr;
-	 //methods that we call in C# if they are available
-	 MonoMethod* m_OnPulseMethod = nullptr;
-	 MonoMethod* m_OnWriteChatColor = nullptr;
-	 MonoMethod* m_OnIncomingChat = nullptr;
-	 MonoMethod* m_OnInit = nullptr;
-	 MonoMethod* m_OnUpdateImGui = nullptr;
-	 MonoMethod* m_OnStop = nullptr;
-	 MonoMethod* m_OnCommand = nullptr;
-	 MonoMethod* m_OnSetSpawns = nullptr;
-	 MonoMethod* m_OnQuery = nullptr;
+#include "MQ2MonoShared.h"
 
-	 std::map<std::string, bool> m_IMGUI_OpenWindows;
-	 std::map<std::string, bool> m_IMGUI_CheckboxValues;
-	 std::map<std::string, std::string> m_IMGUI_InputTextValues;
-	 std::map<std::string, bool> m_IMGUI_RadioButtonValues;
-	 std::string m_CurrentWindow;
-	 bool m_IMGUI_Open = true;
-	 int m_delayTime = 0;//amount of time in milliseonds that was set by C#
-	 std::chrono::steady_clock::time_point m_delayTimer = std::chrono::steady_clock::now(); //the time this was issued + m_delayTime
-	 std::deque<std::string> m_CommandList;
-	 bool AddCommand(std::string commandName)
-	 {
-		 if (!IsCommand(commandName.c_str()))
-		 {
-			 m_CommandList.push_back(commandName);
-
-			 mq::AddCommand(commandName.c_str(), [this, commandName](PlayerClient*, const char* args) -> void
-				 {
-
-					 if (this->m_appDomain)
-					 {
-						 if (this->m_OnCommand)
-						 {
-							 std::string line = args;
-							 line = commandName + " " + line;
-							 mono_domain_set(this->m_appDomain, false);
-							 MonoString* monoLine = mono_string_new(this->m_appDomain, line.c_str());
-							 void* params[1] =
-							 {
-								 monoLine
-
-							 };
-							 mono_runtime_invoke(this->m_OnCommand, this->m_classInstance, params, nullptr);
-						 }
-					 }
-				 });
-			 return true;
-		 }
-		 return false;
-		 
-	 }
-	 void RemoveCommand(std::string command)
-	 {
-		 int count = 0;
-		 int processCount = static_cast<int>(this->m_CommandList.size());
-		 while (count < processCount)
-		 {
-			 count++;
-			 std::string currentKey = this->m_CommandList.front();
-			 m_CommandList.pop_front();
-			 if (!ci_equals(currentKey, command))
-			 {
-				 m_CommandList.push_back(currentKey);
-			 }
-			 else
-			 {
-				 mq::RemoveCommand(currentKey.c_str());
-			 }
-		 }
-	 }
-	 void ClearCommands()
-	 {
-		 while (this->m_CommandList.size() > 0)
-		 {
-			 mq::RemoveCommand(m_CommandList.front().c_str());
-			 m_CommandList.pop_front();
-		 }
-		
-	 }
-
- };
-
- std::map<std::string, monoAppDomainInfo> monoAppDomains;
- std::map<MonoDomain*, std::string> monoAppDomainPtrToString;
+std::map<std::string, monoAppDomainInfo> monoAppDomains;
+std::map<MonoDomain*, std::string> monoAppDomainPtrToString;
  //used to keep a revolving list of who is valid to process. 
  std::deque<std::string> appDomainProcessQueue;
  uint32_t bmUpdateMonoOnPulse = 0;
@@ -408,9 +231,30 @@ void InitMono()
     // Image display
     mono_add_internal_call("MonoCore.Core::imgui_Image", &mono_ImGUI_Image);
 
+    // Fonts
+    mono_add_internal_call("MonoCore.Core::imgui_AddFontFromFileTTF", &mono_ImGUI_AddFontFromFileTTF);
+    mono_add_internal_call("MonoCore.Core::imgui_PushFont", &mono_ImGUI_PushFont);
+    mono_add_internal_call("MonoCore.Core::imgui_PopFont", &mono_ImGUI_PopFont);
+    mono_add_internal_call("MonoCore.Core::imgui_PushMaterialIconsFont", &mono_ImGUI_PushMaterialIconsFont);
+
     // Spell icon drawing
     mono_add_internal_call("MonoCore.Core::imgui_DrawSpellIconByIconIndex", &mono_ImGUI_DrawSpellIconByIconIndex);
     mono_add_internal_call("MonoCore.Core::imgui_DrawSpellIconBySpellID", &mono_ImGUI_DrawSpellIconBySpellID);
+
+    // Drawing functions for custom backgrounds
+    mono_add_internal_call("MonoCore.Core::imgui_GetCursorPosY", &mono_ImGUI_GetCursorPosY);
+    mono_add_internal_call("MonoCore.Core::imgui_GetCursorScreenPosX", &mono_ImGUI_GetCursorScreenPosX);
+    mono_add_internal_call("MonoCore.Core::imgui_GetCursorScreenPosY", &mono_ImGUI_GetCursorScreenPosY);
+    mono_add_internal_call("MonoCore.Core::imgui_GetTextLineHeightWithSpacing", &mono_ImGUI_GetTextLineHeightWithSpacing);
+    mono_add_internal_call("MonoCore.Core::imgui_GetFrameHeight", &mono_ImGUI_GetFrameHeight);
+    mono_add_internal_call("MonoCore.Core::imgui_GetWindowDrawList_AddRectFilled", &mono_ImGUI_GetWindowDrawList_AddRectFilled);
+
+    // Item rect + color helpers
+    mono_add_internal_call("MonoCore.Core::imgui_GetItemRectMinX", &mono_ImGUI_GetItemRectMinX);
+    mono_add_internal_call("MonoCore.Core::imgui_GetItemRectMinY", &mono_ImGUI_GetItemRectMinY);
+    mono_add_internal_call("MonoCore.Core::imgui_GetItemRectMaxX", &mono_ImGUI_GetItemRectMaxX);
+    mono_add_internal_call("MonoCore.Core::imgui_GetItemRectMaxY", &mono_ImGUI_GetItemRectMaxY);
+    mono_add_internal_call("MonoCore.Core::imgui_GetColorU32", &mono_ImGUI_GetColorU32);
 
     // Texture creation from raw data
     mono_add_internal_call("MonoCore.Core::mq_CreateTextureFromData", &mono_CreateTextureFromData);
@@ -1402,595 +1246,7 @@ static void mono_RemoveCommand(MonoString* text)
 		domainInfo.RemoveCommand(str);
 	}
 }
-#pragma region
-//IMGUI test code, not really being used atm as i went into another direction
-static void mono_ImGUI_Begin_OpenFlagSet(MonoString* name, bool open)
-{
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-
-	MonoDomain* currentDomain = mono_domain_get();
-
-	if (currentDomain)
-	{
-		std::string key = monoAppDomainPtrToString[currentDomain];
-		//pointer to the value in the map
-		auto & domainInfo = monoAppDomains[key];
-		if (domainInfo.m_IMGUI_OpenWindows.find(str) == domainInfo.m_IMGUI_OpenWindows.end())
-		{
-			//key doesn't exist, add it
-			domainInfo.m_IMGUI_OpenWindows[str] = true;
-		}
-		domainInfo.m_IMGUI_OpenWindows[str] = open;
-		//put updates back
-	
-	}
-
-	
-
-}
-static boolean mono_ImGUI_Begin_OpenFlagGet(MonoString* name)
-{
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-	MonoDomain* currentDomain = mono_domain_get();
-
-	if (currentDomain)
-	{
-		std::string key = monoAppDomainPtrToString[currentDomain];
-		//pointer to the value in the map
-		auto& domainInfo = monoAppDomains[key];
-		if (domainInfo.m_IMGUI_OpenWindows.find(str) == domainInfo.m_IMGUI_OpenWindows.end())
-		{
-			//key doesn't exist, add it
-			domainInfo.m_IMGUI_OpenWindows[str] = true;
-		}
-		return domainInfo.m_IMGUI_OpenWindows[str];
-	}
-	return false;
-
-}
-//define methods exposde to the plugin to be executed
-static bool mono_ImGUI_Begin(MonoString* name, int flags)
-{
-
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-	MonoDomain* currentDomain = mono_domain_get();
-
-	if (currentDomain)
-	{
-		std::string key = monoAppDomainPtrToString[currentDomain];
-		//pointer to the value in the map
-		auto& domainInfo = monoAppDomains[key];
-
-		domainInfo.m_CurrentWindow = str;
-		if (domainInfo.m_IMGUI_OpenWindows.find(str) == domainInfo.m_IMGUI_OpenWindows.end())
-		{
-			//key doesn't exist, add it
-			domainInfo.m_IMGUI_OpenWindows[str] = true;
-		}
-
-		return ImGui::Begin(str.c_str(), &domainInfo.m_IMGUI_OpenWindows[str], flags);
-	}
-	return false;
-}
-
-
-static bool mono_ImGUI_Button(MonoString* name)
-{
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-	return ImGui::Button(str.c_str());
-}
-
-static bool mono_ImGUI_ButtonEx(MonoString* name, float width, float height)
-{
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-	return ImGui::Button(str.c_str(), ImVec2(width, height));
-}
-
-static bool mono_ImGUI_SmallButton(MonoString* name)
-{
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-	return ImGui::SmallButton(str.c_str());
-}
-
-static void mono_ImGUI_End()
-{
-	ImGui::End();
-}
-
-static void mono_ImGUI_Text(MonoString* text)
-{
-    char* cppString = mono_string_to_utf8(text);
-    std::string str(cppString);
-    mono_free(cppString);
-    ImGui::TextUnformatted(str.c_str());
-}
-
-static void mono_ImGUI_TextUnformatted(MonoString* text)
-{
-    char* cppString = mono_string_to_utf8(text);
-    std::string str(cppString);
-    mono_free(cppString);
-    ImGui::TextUnformatted(str.c_str());
-}
-
-static void mono_ImGUI_Separator()
-{
-	ImGui::Separator();
-}
-
-static void mono_ImGUI_SameLine()
-{
-	ImGui::SameLine();
-}
-
-static void mono_ImGUI_SameLineEx(float offset_from_start_x, float spacing)
-{
-	ImGui::SameLine(offset_from_start_x, spacing);
-}
-
-static bool mono_ImGUI_Checkbox(MonoString* name, bool defaultValue)
-{
-	char* cppString = mono_string_to_utf8(name);
-	std::string str(cppString);
-	mono_free(cppString);
-
-	MonoDomain* currentDomain = mono_domain_get();
-	if (currentDomain)
-	{
-		std::string key = monoAppDomainPtrToString[currentDomain];
-		auto& domainInfo = monoAppDomains[key];
-		bool value = defaultValue;
-		auto it = domainInfo.m_IMGUI_CheckboxValues.find(str);
-		if (it != domainInfo.m_IMGUI_CheckboxValues.end())
-		{
-			value = it->second;
-		}
-		bool changed = ImGui::Checkbox(str.c_str(), &value);
-		domainInfo.m_IMGUI_CheckboxValues[str] = value;
-		return value;
-	}
-	return defaultValue;
-}
-
-static bool mono_ImGUI_BeginTabBar(MonoString* name)
-{
-    char* cppString = mono_string_to_utf8(name);
-    std::string str(cppString);
-    mono_free(cppString);
-    return ImGui::BeginTabBar(str.c_str());
-}
-
-static void mono_ImGUI_EndTabBar()
-{
-    ImGui::EndTabBar();
-}
-
-static bool mono_ImGUI_BeginTabItem(MonoString* label)
-{
-    char* cppString = mono_string_to_utf8(label);
-    std::string str(cppString);
-    mono_free(cppString);
-    return ImGui::BeginTabItem(str.c_str());
-}
-
-static void mono_ImGUI_EndTabItem()
-{
-    ImGui::EndTabItem();
-}
-
-// ===================== ImGui wrappers: Tables =====================
-bool mono_ImGUI_BeginTable(MonoString* id, int columns, int flags, float outer_width)
-{
-    if (!id) return false;
-    char* sid = mono_string_to_utf8(id);
-    bool result = ImGui::BeginTable(sid, columns, (ImGuiTableFlags)flags, ImVec2(outer_width, 0.0f));
-    mono_free(sid);
-    return result;
-}
-void mono_ImGUI_EndTable()
-{
-    ImGui::EndTable();
-}
-void mono_ImGUI_TableSetupColumn(MonoString* label, int flags, float init_width)
-{
-    if (!label) return;
-    char* slabel = mono_string_to_utf8(label);
-    ImGui::TableSetupColumn(slabel, (ImGuiTableColumnFlags)flags, init_width);
-    mono_free(slabel);
-}
-void mono_ImGUI_TableHeadersRow()
-{
-    ImGui::TableHeadersRow();
-}
-void mono_ImGUI_TableNextRow()
-{
-    ImGui::TableNextRow();
-}
-bool mono_ImGUI_TableNextColumn()
-{
-    return ImGui::TableNextColumn();
-}
-
-// ===================== ImGui wrappers: Colors / TextColored =====================
-void mono_ImGUI_TextColored(float r, float g, float b, float a, MonoString* text)
-{
-    if (!text) return;
-    char* stext = mono_string_to_utf8(text);
-    ImGui::TextColored(ImVec4(r, g, b, a), "%s", stext);
-    mono_free(stext);
-}
-void mono_ImGUI_PushStyleColor(int which, float r, float g, float b, float a)
-{
-    ImGui::PushStyleColor((ImGuiCol)which, ImVec4(r, g, b, a));
-}
-void mono_ImGUI_PopStyleColor(int count)
-{
-    ImGui::PopStyleColor(count);
-}
-
-// ===================== ImGui wrappers: Text wrapping and constraints =====================
-void mono_ImGUI_TextWrapped(MonoString* text)
-{
-    if (!text) return;
-    char* stext = mono_string_to_utf8(text);
-    ImGui::TextWrapped("%s", stext);
-    mono_free(stext);
-}
-void mono_ImGUI_PushTextWrapPos(float wrap_local_pos_x)
-{
-    ImGui::PushTextWrapPos(wrap_local_pos_x);
-}
-void mono_ImGUI_PopTextWrapPos()
-{
-    ImGui::PopTextWrapPos();
-}
-void mono_ImGUI_SetNextWindowSizeConstraints(float min_w, float min_h, float max_w, float max_h)
-{
-    ImGui::SetNextWindowSizeConstraints(ImVec2(min_w, min_h), ImVec2(max_w, max_h));
-}
-
-static void mono_ImGUI_SetNextWindowBgAlpha(float alpha)
-{
-    ImGui::SetNextWindowBgAlpha(alpha);
-}
-
-static void mono_ImGUI_SetNextWindowSize(float width, float height)
-{
-    ImGui::SetNextWindowSize(ImVec2(width, height));
-}
-
-static bool mono_ImGUI_IsWindowHovered()
-{
-    return ImGui::IsWindowHovered();
-}
-
-static bool mono_ImGUI_IsMouseClicked(int button)
-{
-    return ImGui::IsMouseClicked(button);
-}
-
-static bool mono_ImGUI_BeginChild(MonoString* id, float width, float height, bool border)
-{
-    char* cppString = mono_string_to_utf8(id);
-    std::string str(cppString);
-    mono_free(cppString);
-    ImVec2 size(width, height);
-    return ImGui::BeginChild(str.c_str(), size, border);
-}
-
-static void mono_ImGUI_EndChild()
-{
-    ImGui::EndChild();
-}
-
-static bool mono_ImGUI_Selectable(MonoString* label, bool selected)
-{
-    char* cppString = mono_string_to_utf8(label);
-    std::string str(cppString);
-    mono_free(cppString);
-    return ImGui::Selectable(str.c_str(), selected);
-}
-
-static float mono_ImGUI_GetContentRegionAvailX()
-{
-    return ImGui::GetContentRegionAvail().x;
-}
-
-static float mono_ImGUI_GetContentRegionAvailY()
-{
-    return ImGui::GetContentRegionAvail().y;
-}
-
-static bool mono_ImGUI_BeginCombo(MonoString* label, MonoString* preview, int flags)
-{
-    char* lC = mono_string_to_utf8(label);
-    std::string lStr(lC);
-    mono_free(lC);
-    char* pC = mono_string_to_utf8(preview);
-    std::string pStr(pC ? pC : "");
-    if (pC) mono_free(pC);
-    return ImGui::BeginCombo(lStr.c_str(), pStr.c_str(), (ImGuiComboFlags)flags);
-}
-
-static void mono_ImGUI_EndCombo()
-{
-    ImGui::EndCombo();
-}
-
-static bool mono_ImGUI_RightAlignButton(MonoString* name)
-{
-    if (!name) return false;
-    char* cpp = mono_string_to_utf8(name);
-    std::string label(cpp ? cpp : "");
-    if (cpp) mono_free(cpp);
-
-    ImVec2 textSize = ImGui::CalcTextSize(label.c_str());
-    float buttonWidth = textSize.x + ImGui::GetStyle().FramePadding.x * 2.0f;
-    float avail = ImGui::GetContentRegionAvail().x;
-    if (avail > buttonWidth)
-    {
-        ImGui::SetCursorPosX(ImGui::GetCursorPosX() + (avail - buttonWidth));
-    }
-    return ImGui::Button(label.c_str());
-}
-
-static bool mono_ImGUI_InputText(MonoString* id, MonoString* initial)
-{
-    char* idC = mono_string_to_utf8(id);
-    std::string idStr(idC);
-    mono_free(idC);
-
-    char* initC = mono_string_to_utf8(initial);
-    std::string initStr(initC ? initC : "");
-    if (initC) mono_free(initC);
-
-    MonoDomain* currentDomain = mono_domain_get();
-    if (!currentDomain) return false;
-
-    std::string key = monoAppDomainPtrToString[currentDomain];
-    auto& domainInfo = monoAppDomains[key];
-    auto it = domainInfo.m_IMGUI_InputTextValues.find(idStr);
-    if (it == domainInfo.m_IMGUI_InputTextValues.end())
-    {
-        domainInfo.m_IMGUI_InputTextValues[idStr] = initStr;
-        it = domainInfo.m_IMGUI_InputTextValues.find(idStr);
-    }
-
-    // limit buffer to 512 chars
-    char buf[512];
-    memset(buf, 0, sizeof(buf));
-    strncpy(buf, it->second.c_str(), sizeof(buf) - 1);
-    bool changed = ImGui::InputText(idStr.c_str(), buf, sizeof(buf));
-    if (changed)
-    {
-        it->second = std::string(buf);
-        return true;
-    }
-    return false;
-}
-
-static MonoString* mono_ImGUI_InputText_Get(MonoString* id)
-{
-    char* idC = mono_string_to_utf8(id);
-    std::string idStr(idC);
-    mono_free(idC);
-
-    MonoDomain* currentDomain = mono_domain_get();
-    if (!currentDomain) return mono_string_new(mono_get_root_domain(), "");
-    std::string key = monoAppDomainPtrToString[currentDomain];
-    auto& domainInfo = monoAppDomains[key];
-    auto it = domainInfo.m_IMGUI_InputTextValues.find(idStr);
-    if (it == domainInfo.m_IMGUI_InputTextValues.end())
-    {
-        return mono_string_new(currentDomain, "");
-    }
-    return mono_string_new(currentDomain, it->second.c_str());
-}
-
-static void mono_ImGUI_SetNextItemWidth(float width)
-{
-    ImGui::SetNextItemWidth(width);
-}
-
-// ----- Slider helpers -----
-static bool mono_ImGUI_SliderInt(MonoString* id, int* value, int min, int max)
-{
-    if (!id || !value) return false;
-    char* idC = mono_string_to_utf8(id);
-    std::string idStr(idC ? idC : "");
-    if (idC) mono_free(idC);
-
-    int v = *value;
-    bool changed = ImGui::SliderInt(idStr.c_str(), &v, min, max);
-    if (changed) { *value = v; }
-    return changed;
-}
-
-static bool mono_ImGUI_SliderDouble(MonoString* id, double* value, double min, double max, MonoString* fmt)
-{
-    if (!id || !value) return false;
-    char* idC = mono_string_to_utf8(id);
-    std::string idStr(idC ? idC : "");
-    if (idC) mono_free(idC);
-
-    std::string fmtStr = "%.3f";
-    if (fmt)
-    {
-        char* f = mono_string_to_utf8(fmt);
-        if (f) { fmtStr = f; mono_free(f); }
-    }
-
-    double v = *value;
-    bool changed = ImGui::SliderScalar(idStr.c_str(), ImGuiDataType_Double, &v, &min, &max, fmtStr.c_str());
-    if (changed) { *value = v; }
-    return changed;
-}
-
-// ----- Context menu / popup helpers -----
-static bool mono_ImGUI_BeginPopupContextItem(MonoString* id, int flags)
-{
-    const char* c_id = nullptr;
-    std::string temp;
-    if (id)
-    {
-        char* s = mono_string_to_utf8(id);
-        temp = (s ? s : "");
-        if (s) mono_free(s);
-        c_id = temp.c_str();
-    }
-    return ImGui::BeginPopupContextItem(c_id, (ImGuiPopupFlags)flags);
-}
-
-static bool mono_ImGUI_BeginPopupContextWindow(MonoString* id, int flags)
-{
-    const char* c_id = nullptr;
-    std::string temp;
-    if (id)
-    {
-        char* s = mono_string_to_utf8(id);
-        temp = (s ? s : "");
-        if (s) mono_free(s);
-        c_id = temp.c_str();
-    }
-    return ImGui::BeginPopupContextWindow(c_id, (ImGuiPopupFlags)flags);
-}
-
-static void mono_ImGUI_EndPopup()
-{
-    ImGui::EndPopup();
-}
-
-static bool mono_ImGUI_MenuItem(MonoString* label)
-{
-    if (!label) return false;
-    char* l = mono_string_to_utf8(label);
-    std::string s(l ? l : "");
-    if (l) mono_free(l);
-    return ImGui::MenuItem(s.c_str());
-}
-
-// ----- TreeNode helpers -----
-static bool mono_ImGUI_TreeNode(MonoString* label)
-{
-    if (!label) return false;
-    char* l = mono_string_to_utf8(label);
-    std::string s(l ? l : "");
-    if (l) mono_free(l);
-    return ImGui::TreeNode(s.c_str());
-}
-
-static bool mono_ImGUI_TreeNodeEx(MonoString* label, int flags)
-{
-    if (!label) return false;
-    char* l = mono_string_to_utf8(label);
-    std::string s(l ? l : "");
-    if (l) mono_free(l);
-    return ImGui::TreeNodeEx(s.c_str(), (ImGuiTreeNodeFlags)flags);
-}
-
-static void mono_ImGUI_TreePop()
-{
-    ImGui::TreePop();
-}
-
-static bool mono_ImGUI_CollapsingHeader(MonoString* label, int flags)
-{
-    if (!label) return false;
-    char* l = mono_string_to_utf8(label);
-    std::string s(l ? l : "");
-    if (l) mono_free(l);
-    return ImGui::CollapsingHeader(s.c_str(), (ImGuiTreeNodeFlags)flags);
-}
-
-// ----- Tooltips and hover detection -----
-static bool mono_ImGUI_IsItemHovered()
-{
-    return ImGui::IsItemHovered();
-}
-
-static void mono_ImGUI_BeginTooltip()
-{
-    ImGui::BeginTooltip();
-}
-
-static void mono_ImGUI_EndTooltip()
-{
-    ImGui::EndTooltip();
-}
-
-// ----- Image display -----
-static void mono_ImGUI_Image(void* textureId, float width, float height)
-{
-    ImGui::Image(textureId, ImVec2(width, height));
-}
-
-// ----- Spell icon drawing -----
-static void mono_ImGUI_DrawSpellIconByIconIndex(int iconIndex, float size)
-{
-    static CTextureAnimation* s_pTASpellIcons = nullptr;
-
-    if (!pSidlMgr)
-        return;
-
-    if (!s_pTASpellIcons)
-    {
-        if (CTextureAnimation* temp = pSidlMgr->FindAnimation("A_SpellGems"))
-        {
-            s_pTASpellIcons = new CTextureAnimation(*temp);
-        }
-    }
-
-    if (!s_pTASpellIcons)
-        return;
-
-    if (iconIndex < 0)
-        return;
-
-    s_pTASpellIcons->SetCurCell(iconIndex);
-    mq::imgui::DrawTextureAnimation(s_pTASpellIcons, eqlib::CXSize((int)size, (int)size));
-}
-
-static void mono_ImGUI_DrawSpellIconBySpellID(int spellId, float size)
-{
-    if (spellId <= 0)
-        return;
-
-    eqlib::EQ_Spell* pSpell = GetSpellByID(spellId);
-    if (!pSpell)
-        return;
-
-    int iconIndex = pSpell->SpellIcon;
-    mono_ImGUI_DrawSpellIconByIconIndex(iconIndex, size);
-}
-
-// ----- Texture creation from raw data -----
-static void* mono_CreateTextureFromData(const uint8_t* data, int width, int height, int channels)
-{
-    // For now, return nullptr as a placeholder
-    // A full implementation would need to create a DirectX/OpenGL texture from the raw data
-    // and return an ImTextureID that ImGui can use
-    return nullptr;
-}
-
-static void mono_DestroyTexture(void* textureId)
-{
-    // Placeholder for texture cleanup
-    // A full implementation would free the DirectX/OpenGL texture resources
-}
-
-#pragma endregion test code for ImGUI, went into another direction
+#pragma endregion
 
 static void mono_Delay(int milliseconds)
 {
