+++ MQ2Mono.vcxproj
@@ -99,9 +99,12 @@
   </PropertyGroup>
   <ItemGroup>
     <ClCompile Include="MQ2Mono.cpp" />
+    <ClCompile Include="MQ2MonoImGui.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h" />
+    <ClInclude Include="MQ2MonoImGui.h" />
+    <ClInclude Include="MQ2MonoShared.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="MQ2Mono.rc" />
